+++ ReleaseNotes.md
@@ -1,5 +1,9 @@
 # Release Notes for PoshSemanticVersion
 
-## 1.1.0 - 2017-10-30
+## 1.2.0 - 2017-11-01
 ### Features
-- Step-SemanticVersion can increment prerelease and build using optional label parameter.
+- Test-SemanticVersion provides useful feedback messages if a Semantic Version is invalid and the -Verbose switch
+  is used.
+
+### Fixes
+- Test-SemanticVersion correctly accepts pipeline input of more than one string.